Move reshaped_ functions to _tensors.py
