+++ phiml/math/__init__.py
@@ -32,7 +32,7 @@ from ._magic_ops import (
     copy_with, replace, find_differences
 )
 
-from ._tensors import wrap, tensor, layout, Tensor, Dict, to_dict, from_dict, is_scalar, BROADCAST_FORMATTER as f
+from ._tensors import Tensor, wrap, tensor, layout, native, numpy_ as numpy, reshaped_numpy, Dict, to_dict, from_dict, is_scalar, BROADCAST_FORMATTER as f
 
 from ._sparse import dense, get_sparsity, get_format, to_format, is_sparse, sparse_tensor, stored_indices, stored_values, tensor_like
 
@@ -40,7 +40,7 @@ from .extrapolation import Extrapolation, as_extrapolation
 
 from ._ops import (
     choose_backend_t as choose_backend, all_available, convert, seed, to_device,
-    native, numpy, reshaped_native, reshaped_tensor, reshaped_numpy, copy, native_call,
+    reshaped_native, reshaped_tensor, copy, native_call,
     print_ as print,
     slice_off,
     zeros, ones, fftfreq, random_normal, random_uniform, meshgrid, linspace, arange as range, range_tensor,  # creation operators (use default backend)
